requirements.txt
PyYAML
python-can
pyserial
RPi.GPIO
smbus2